refactor(metadata): Delete server/src/metadata/model.rs
