+++ server/src/metadata/mod.rs
@@ -1,4 +1,3 @@
 pub use messages::MetadataMessage;
 
 mod messages;
-mod model;
